requirements.txt
Flask==3.0.0
pdf2image==1.16.3
Pillow==10.1.0
python-pptx==0.6.23
PyMuPDF==1.23.8
gunicorn==21.2.0

